pretrain_ablations/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "pretrain_ablations"
version = "0.1.0"
description = "Modular pretraining ablation pipeline for IMU behavior recognition"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.1",
    "transformers>=4.40.0",
    "pandas>=2.0",
    "pyarrow",
    "numpy",
    "scikit-learn",
    "matplotlib",
    "seaborn",
    "omegaconf",
    "pyyaml",
    "tqdm",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
